Implement an effect bar to display effect duration
This is a rather large commit but it resolves a lot of minor things so
that effects will now display a bar with the remaining duration of the
effect. The idea is to stack based on number of effects and if it is the
same effect perhaps even a number?