Refactored textureID handling and effect rendering
I initially thought that it would be wise for each class to decide for
textureIDs themselves but I quickly realised that this could make
definitions confusing. For example I accidentally set both the berry and
vinjette to be 0x0 which broke parts of the game.

In this change I also included a way for effects to be rendered. The
first effect to be rendered is the glowing effect for the golden berry!